requirements.txt
#the following are for Linux and Mac
pillows
transformers
torch
einops
pyvips-binary
pyvips